update game.js api fetch data
