+++ src/style.scss
@@ -22,7 +22,7 @@ body {
   }
 }
 
-#leaderboard-container {
+#leaderboard {
   margin-top: 30px;
   display: flex;
   flex-direction: column;
@@ -67,11 +67,11 @@ body {
   }
 }
 
-#scores-container {
-  padding: 22px;
+#all-scores {
+  padding: 20px;
   display: flex;
   flex-direction: column;
-  gap: 22px;
+  gap: 20px;
 
   #header {
     display: flex;
@@ -160,18 +160,18 @@ body {
     overflow-y: auto;
   }
 
-  #leaderboard-container {
+  #leaderboard {
     flex-direction: row;
     justify-content: center;
     align-items: flex-start;
     margin: auto;
 
-    & > * {
+    &>* {
       width: 50%;
       max-width: 550px;
     }
 
-    #scores-container > ul {
+    #all-scores>ul {
       background-color: #fff;
     }
   }
